requirements.txt
Django==1.9.1
django-crispy-forms==1.6.0
